optimize main page
